Major internals refactoring, change data model, change storage, change UI API. Intermediate step

+++ money-keeper-core/src/main/java/com/github/money/keeper/storage/AccountRepo.java
@@ -1,6 +1,6 @@
 package com.github.money.keeper.storage;
 
-import com.github.money.keeper.model.Account;
+import com.github.money.keeper.model.core.Account;
 
-public interface AccountRepo extends BaseRepo<Integer, Account> {
+public interface AccountRepo extends BaseRepo<Long, Account> {
 }
